+++ src/app/profile/view/(components)/organizationEdit2.test.js
@@ -9,6 +9,7 @@ import "@testing-library/jest-dom";
 
 import React from "react";
 import Organization from "./organization";
+import { orgSchema } from "./organization";
 import { describe } from "node:test";
 
 afterEach(() => {
@@ -64,53 +65,6 @@ describe("Organization Profile View/Edit Tests", () => {
     [[], null],
   ];
   //Test 1
-  test("Organization Page Rendering all elements", async () => {
-    const setEditProfile = jest.fn();
-    render(
-      <Organization
-        organizations={organizations}
-        listings={listings}
-        editProfile={false}
-        setEditProfile={setEditProfile}
-      />
-    );
-    const select_button = await screen.findByText("Select");
-    const name = await screen.findByText("Company Name:");
-    const email = await screen.findByText("Email Address:");
-    const bio = await screen.findByText("Biography:");
-    const address = await screen.findByText("Address:");
-    const phone = await screen.findByText("Phone:");
-    const edit_button = await screen.findByText("Edit");
-
-    expect(select_button).toBeInTheDocument();
-    expect(name).toBeInTheDocument();
-    expect(email).toBeInTheDocument();
-    expect(bio).toBeInTheDocument();
-    expect(address).toBeInTheDocument();
-    expect(phone).toBeInTheDocument();
-    expect(edit_button).toBeInTheDocument();
-  });
-
-  //Test 2
-  test("Organization Edit Button working", async () => {
-    const setEditProfile = jest.fn();
-    render(
-      <Organization
-        organizations={organizations}
-        listings={listings}
-        editProfile={false}
-        setEditProfile={setEditProfile}
-      />
-    );
-
-    const edit_button = await screen.findByText("Edit");
-    expect(edit_button).toBeInTheDocument();
-
-    await fireEvent.click(edit_button);
-    expect(setEditProfile).toHaveBeenCalled();
-  });
-
-  //Test 3
   test("Organization Edit Page Rendering all elements", async () => {
     const setEditProfile = jest.fn();
     render(
@@ -140,7 +94,7 @@ describe("Organization Profile View/Edit Tests", () => {
     expect(submit_button).toBeInTheDocument();
     expect(cancel_button).toBeInTheDocument();
   });
-  //Test 4
+  //Test 2
   test("Toggle Edit picture button working", async () => {
     const setEditProfile = jest.fn();
     render(
@@ -161,7 +115,7 @@ describe("Organization Profile View/Edit Tests", () => {
 
     expect(image_upload).toBeInTheDocument();
   });
-  //Test 5
+  //Test 3
   test("Button Testing", async () => {
     const setEditProfile = jest.fn();
     render(
@@ -182,26 +136,78 @@ describe("Organization Profile View/Edit Tests", () => {
     expect(setEditProfile).toHaveBeenCalled();
   });
 
-  //Test 6
-  test("Button Testing 2", async () => {
-    const setEditProfile = jest.fn();
-    render(
-      <Organization
-        organizations={organizations}
-        listings={listings}
-        editProfile={true}
-        setEditProfile={setEditProfile}
-      />
-    );
-    const submit_button = await screen.findByTestId("submit");
+  //Test 4
+  test("Schema Testing Success", async () => {
+    const test = orgSchema.safeParse({
+      name: "Test Org2",
+      email: "test@test.com",
+      address: "4576 Giinagay Way, Raleigh New South Wales 2455, Australia",
+      phoneNumber: "13456778900",
+      bio: "Testing 123",
+    });
 
-    console.log(submit_button);
-    expect(submit_button).toBeInTheDocument();
-    expect(submit_button).not.toBeDisabled();
-    expect(submit_button).toBeVisible();
-
-    fireEvent.click(submit_button);
-
-    expect(setEditProfile).toHaveBeenCalled();
+    expect(test.success).toBe(true);
   });
+  //Test 5
+  test("Schema Testing Email Fail", async () => {
+    const test = orgSchema.safeParse({
+      name: "Test Org2",
+      email: "test@test",
+      address: "4576 Giinagay Way, Raleigh New South Wales 2455, Australia",
+      phoneNumber: "13456778900",
+      bio: "Testing 123",
+    });
+
+    expect(test.success).toBe(false);
+  });
+  // Test 6
+  test("Schema Testing Name Fail", async () => {
+    const test = orgSchema.safeParse({
+      name: "",
+      email: "test@test.com",
+      address: "4576 Giinagay Way, Raleigh New South Wales 2455, Australia",
+      phoneNumber: "13456778900",
+      bio: "Testing 123",
+    });
+
+    expect(test.success).toBe(false);
+  });
+  // Test 7
+  test("Schema Testing Address Fail", async () => {
+    const test = orgSchema.safeParse({
+      name: "Test Org2",
+      email: "test@test.com",
+      address: "",
+      phoneNumber: "13456778900",
+      bio: "Testing 123",
+    });
+
+    expect(test.success).toBe(false);
+  });
+  // Test 8
+  test("Schema Testing Phonenumber fail", async () => {
+    const test = orgSchema.safeParse({
+      name: "Test Org2",
+      email: "test@test.com",
+      address: "4576 Giinagay Way, Raleigh New South Wales 2455, Australia",
+      phoneNumber: "",
+      bio: "Testing 123",
+    });
+
+    expect(test.success).toBe(false);
+  });
+  // Test 9
+  test("Schema Testing bio fail", async () => {
+    const test = orgSchema.safeParse({
+      name: "Test Org2",
+      email: "test@test.com",
+      address: "4576 Giinagay Way, Raleigh New South Wales 2455, Australia",
+      phoneNumber: "13456778900",
+      bio: "",
+    });
+
+    expect(test.success).toBe(false);
+  });
+
+  //Did not test Submit button because I believe it is being blocked by zod field checks
 });
